refactor: move timer decorator to a new module for better organization
