+++ src/imgtools/ops/input_classes.py
@@ -23,36 +23,10 @@ from imgtools.logging import logger
 from imgtools.modules.datagraph import DataGraph
 
 from imgtools.ops.base_classes import BaseInput
+from imgtools.utils.timer import timer
 
 LoaderFunction = Callable[..., sitk.Image | StructureSet | Segmentation]
 
-
-def timer(name: str) -> Callable[[Callable[..., Any]], Callable[..., Any]]:
-    """
-    Decorator to measure the execution time of a function and log it with a custom name.
-
-    Parameters
-    ----------
-        name (str): The custom name to use in the log message.
-
-    Returns
-    -------
-        Callable[[Callable[..., Any]], Callable[..., Any]]:
-        A decorator that wraps the function to measure its execution time.
-    """
-
-    def decorator(func: Callable[..., Any]) -> Callable[..., Any]:
-        def wrapper(*args: Any, **kwargs: Any) -> Any:
-            start_time = time.time()
-            result = func(*args, **kwargs)
-            end_time = time.time()
-            elapsed_time = end_time - start_time
-            logger.info(f"{name} took {elapsed_time:.4f} seconds")
-            return result
-
-        return wrapper
-
-    return decorator
 
 class ImageMask(NamedTuple):
     """
